Create settings for generating prompts

Create settings for generating prompts

createPromptSettings(
  writeNarrative = TRUE,
  testingReminder = TRUE,
  uncertaintyInstructions = TRUE,
  discussEvidence = TRUE,
  provideExamples = FALSE,
  maxParts = 100,
  maxDays = 5
)

Arguments

writeNarrative

Ask the LLM to write a clinical narrative matching the provided data?

testingReminder

Remind the LLM that a diagnosis can be recorded just to justify a test, and therefore by itself is not sufficient evidence?

uncertaintyInstructions

Provide instructions to the LLM on how to deal with uncertainty?

discussEvidence

Prompt the LLM to first discuss evidence in favor and against the disease of interest?

provideExamples

Provide examples? (few-shot prompting)

maxParts

How many parts can a category have? For example, if maxParts = 100 and there are more than 100 measurements, a random sample of 100 will be taken. Set to 0 if there is no maximum.

maxDays

How many days can a single code have? For example, if maxDays = 5 and there is a measurement code that appears on more than 5 days, a random sample of 5 days will be taken. Set to 0 if there is no maximum.

Value

A settings object, to be used in createSystemPrompt() and createPrompt().